\name{MFIaggr.gui}
\alias{MFIaggr.gui}

\title{
Multiple Comparison of the Cycle Dependent Variance - Graphical User Interface
}
\description{
Blabla
}
\usage{
MFIaggr.gui()
}

\author{
Stefan Roediger, Michal Burdukiewicz.
}
\section{Warning }{
Any ad-blocking software may be cause of malfunctions. 
}

\examples{
\donttest{
# Run from a R console following commands
require(shiny)

# Invoke the shiny AmpSim app in the default browser
runApp(paste0(find.package("chipPCR")[1],"/MFIaggr.gui"))
}
}

\seealso{
\code{\link[chipPCR]{MFIaggr}}
}

\keyword{ hplot }
\keyword{ simulation }
\keyword{ shiny }
\keyword{ GUI }
\keyword{ browser }
