\name{AmpSim.gui}
\alias{AmpSim.gui}

\title{
Amplification Curve Simulation Graphical User Interface
}
\description{
Launches graphical user interface that allows simulating and analysing amplification reactions.
}
\usage{
AmpSim.gui()
}

\author{
Stefan Roediger, Michal Burdukiewicz.
}
\section{Warning }{
Any ad-blocking software may be cause of malfunctions. 
}

\seealso{
\code{\link[chipPCR]{AmpSim}}, \code{\link[chipPCR]{bg.max}}
}

\keyword{ hplot }
