\name{MFIaggr.gui}
\alias{MFIaggr.gui}

\title{
Multiple Comparison of the Cycle Dependent Variance - Graphical User 
Interface
}
\description{
\code{MFIaggr.gui} is a Graphical User Interface based on the \code{shiny} 
package. The core if this GUI is based on the Comparison of the 
\code{\link[chipPCR]{MFIaggr}} function. The cycle dependent variance can
be analyzed intuitively. This GUI dynamically generates reports.
}
\usage{
MFIaggr.gui()
}

\author{
Stefan Roediger, Michal Burdukiewicz.
}
\section{Warning }{
Any ad-blocking software may be cause of malfunctions. 
}



\seealso{
\code{\link[chipPCR]{MFIaggr}}
}

\keyword{ hplot }
\keyword{ shiny }
\keyword{ GUI }
\keyword{ browser }
