\name{summary-refMFI}
\alias{summary.refMFI}
\alias{summary,refMFI-method}

\title{
Summary \code{refMFI} objects
}
\description{
Summarize \code{\linkS4class{refMFI}} objects and access FIX ME.
}
\usage{
\S4method{summary}{refMFI}(object, print = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
is a \code{\linkS4class{refMFI}} object.
}
  \item{print}{
is a \code{logical} value determining if pretty summary of the object should be printed.
}
 
}


\author{
Stefan Roediger, Michal Burdukiewicz
}

\seealso{
\code{\linkS4class{refMFI}}
}
\examples{

#PUT STH HERE
}
\keyword{manip}